Move tests from src into tests dir
