+++ src/history.rs
@@ -378,89 +378,3 @@ impl<E> Branch<E> {
         Branch { parent, entries }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::*;
-
-    #[test]
-    fn go_to() {
-        //          m
-        //          |
-        //    j  k  l
-        //     \ | /
-        //       i
-        //       |
-        // e  g  h
-        // |  | /
-        // d  f  p - q *
-        // | /  /
-        // c  n - o
-        // | /
-        // b
-        // |
-        // a
-        let mut target = String::new();
-        let mut history = History::new();
-        history.edit(&mut target, Add('a'));
-        history.edit(&mut target, Add('b'));
-        history.edit(&mut target, Add('c'));
-        history.edit(&mut target, Add('d'));
-        history.edit(&mut target, Add('e'));
-        assert_eq!(target, "abcde");
-        history.undo(&mut target).unwrap();
-        history.undo(&mut target).unwrap();
-        assert_eq!(target, "abc");
-        let abc = history.head();
-
-        history.edit(&mut target, Add('f'));
-        history.edit(&mut target, Add('g'));
-        assert_eq!(target, "abcfg");
-        let abcfg = history.head();
-
-        history.undo(&mut target).unwrap();
-        history.edit(&mut target, Add('h'));
-        history.edit(&mut target, Add('i'));
-        history.edit(&mut target, Add('j'));
-        assert_eq!(target, "abcfhij");
-        let abcfhij = history.head();
-
-        history.undo(&mut target).unwrap();
-        history.edit(&mut target, Add('k'));
-        assert_eq!(target, "abcfhik");
-        let abcfhik = history.head();
-
-        history.undo(&mut target).unwrap();
-        history.edit(&mut target, Add('l'));
-        assert_eq!(target, "abcfhil");
-        history.edit(&mut target, Add('m'));
-        assert_eq!(target, "abcfhilm");
-        let abcfhilm = history.head();
-        history.go_to(&mut target, At::new(abc.root, 2));
-        history.edit(&mut target, Add('n'));
-        history.edit(&mut target, Add('o'));
-        assert_eq!(target, "abno");
-        let abno = history.head();
-
-        history.undo(&mut target).unwrap();
-        history.edit(&mut target, Add('p'));
-        history.edit(&mut target, Add('q'));
-        assert_eq!(target, "abnpq");
-
-        let abnpq = history.head();
-        history.go_to(&mut target, abc);
-        assert_eq!(target, "abc");
-        history.go_to(&mut target, abcfg);
-        assert_eq!(target, "abcfg");
-        history.go_to(&mut target, abcfhij);
-        assert_eq!(target, "abcfhij");
-        history.go_to(&mut target, abcfhik);
-        assert_eq!(target, "abcfhik");
-        history.go_to(&mut target, abcfhilm);
-        assert_eq!(target, "abcfhilm");
-        history.go_to(&mut target, abno);
-        assert_eq!(target, "abno");
-        history.go_to(&mut target, abnpq);
-        assert_eq!(target, "abnpq");
-    }
-}
