+++ src/record.rs
@@ -379,52 +379,3 @@ impl<E, F> From<History<E, F>> for Record<E, F> {
         history.record
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::*;
-    use alloc::string::String;
-    use alloc::vec::Vec;
-
-    #[test]
-    fn go_to() {
-        let mut target = String::new();
-        let mut record = Record::new();
-        record.edit(&mut target, Add('a'));
-        record.edit(&mut target, Add('b'));
-        record.edit(&mut target, Add('c'));
-        record.edit(&mut target, Add('d'));
-        record.edit(&mut target, Add('e'));
-
-        record.go_to(&mut target, 0);
-        assert_eq!(record.head(), 0);
-        assert_eq!(target, "");
-        record.go_to(&mut target, 5);
-        assert_eq!(record.head(), 5);
-        assert_eq!(target, "abcde");
-        record.go_to(&mut target, 1);
-        assert_eq!(record.head(), 1);
-        assert_eq!(target, "a");
-        record.go_to(&mut target, 4);
-        assert_eq!(record.head(), 4);
-        assert_eq!(target, "abcd");
-        record.go_to(&mut target, 2);
-        assert_eq!(record.head(), 2);
-        assert_eq!(target, "ab");
-        record.go_to(&mut target, 3);
-        assert_eq!(record.head(), 3);
-        assert_eq!(target, "abc");
-        assert!(record.go_to(&mut target, 6).is_empty());
-        assert_eq!(record.head(), 3);
-    }
-
-    #[test]
-    fn edits() {
-        let mut target = String::new();
-        let mut record = Record::new();
-        record.edit(&mut target, Add('a'));
-        record.edit(&mut target, Add('b'));
-        let collected = record.edits().collect::<Vec<_>>();
-        assert_eq!(&collected[..], &[&Add('a'), &Add('b')][..]);
-    }
-}
